% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{addAttribute}
\alias{addAttribute}
\title{Function to add an attribute class object to a \code{\linkS4class{Dataset}}}
\usage{
addAttribute(data, attr)
}
\arguments{
\item{data}{Object of class \code{\linkS4class{Dataset}}}

\item{attr}{Object of class \code{\linkS4class{Attribute}} or vector to add to the dataset}
}
\value{
The object of class \code{\linkS4class{Dataset}} with a new object of class \code{\linkS4class{Attribute}} on its' list
}
\description{
This adds an attribute to a \code{\linkS4class{Dataset}}
}
\examples{
data<-dataset(c(attribute(c(2,3,6,4)),attribute(c(2,7,8,9))))
addAttribute(data,c(3.2,4.6,9.7,3.2))
attr<-attribute(c(3,73,25,17),name="age")
addAttribute(data,attr)
}
