% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{asMatrix}
\alias{asMatrix}
\title{Function to convert the \code{\linkS4class{DataSet}} into a Matrix}
\usage{
asMatrix(x)
}
\arguments{
\item{x}{\code{\linkS4class{DataSet}} class vector}
}
\value{
A matrix with the data of the \code{\linkS4class{DataSet}}.
}
\description{
This function returns a Matrix with the data of a given \code{\linkS4class{DataSet}}
}
\examples{

}
