% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\docType{class}
\name{Attribute-class}
\alias{Attribute-class}
\title{An S4 class to represent attributes}
\description{
An S4 class to represent attributes
}
\section{Slots}{

\describe{
\item{\code{type}}{Character value indicating the type of the vector: double, factor or numeric}

\item{\code{size}}{Numeric value indicating the size of the vector}

\item{\code{vector}}{Numeric vector containing the information of the \code{\linkS4class{Attribute}}}

\item{\code{name}}{Character with the name of the \code{\linkS4class{Attribute}}}
}}

