% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{attribute}
\alias{attribute}
\title{Basic constructor of the Attribute class}
\usage{
attribute(vector, name, discretize, discretizationType, num.bins)
}
\arguments{
\item{vector}{Numeric vector containing the information of the attribute}

\item{name}{Character indicating the name given to the attribute}

\item{discretize}{Logical optional value indicating if the attribute has to be discretized}

\item{discretizationType}{Character optional value "EW" indicates to compute equal width discretization. "EF" indicates to compute equal frequency discretization}

\item{num.bins}{Numeric number of intervals for discretization}
}
\value{
An object of class \code{\linkS4class{Attribute}} that represents the attribute passed as an argument
}
\description{
This function creates an object of class \code{\linkS4class{Attribute}}
}
\examples{
attribute(c(2,56,87,32,15))
attribute(c(0,1,1,0,1,0))

}
