% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeCorrelation}
\alias{computeCorrelation}
\title{Function to compute the correlation between two objects of class \code{\linkS4class{Attribute}}}
\usage{
computeCorrelation(x, y, discretizationType, num.bins)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{Attribute}}}

\item{y}{an object of class \code{\linkS4class{Attribute}}}

\item{discretizationType}{if x and y are a mix of continuous and discrete variables a discretization is computed
in the continuous one. This parameter indicates the discretization type to compute: Equal Width "EW"
(default) or Equal Frequency "EF"}

\item{num.bins}{if x and y are a mix of continuous and discrete variables a discretization is computed
in the continuous one. This parameter indicates the number of intervals to use in the discretization. The default value is 3.}
}
\value{
A real number with the correlation between both \code{\linkS4class{Attribute}}
}
\description{
This function computes the correlation between two objects of class \code{\linkS4class{Attribute}}
}
\examples{
attr1<-attribute(c(2.3,4.5,6.7,8.9))
attr2<-attribute(c(2,2,1,3))
computeCorrelation(attr1,attr2)
computeCorrelation(attr1,attr2,"EF")
computeCorrelation(attr1,attr2,"EF",4)

}
