% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computEntropy}
\alias{computEntropy}
\title{Function to compute the entropy of a given vector}
\usage{
computEntropy(x, normalize)
}
\arguments{
\item{x}{a vector composed by discrete variables}

\item{normaliza}{a boolean to tell if the function has to be normalized}
}
\value{
A real number
}
\description{
This function computes the entropy of a given vector
}
\examples{
computEntropy(c(2,3,4,5,6,7,7),FALSE)
computEntropy(c(2,3,4,5,6,7,7),TRUE)

}
