% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{correlation}
\alias{correlation}
\title{Function to compute the correlation matrix between the \code{\linkS4class{Attribute}} pairs of the \code{\linkS4class{DataSet}}}
\usage{
correlation(x, discretizationType, num.bins)
}
\arguments{
\item{x}{\code{\linkS4class{DataSet}} class vector}

\item{discretizationType}{if x and y are a mix of continuous and discrete variables a discretization is computed
in the continuous one. This parameter indicates the discretization type to compute: Equal Width "EW"
(default) or Equal Frequency "EF"}

\item{num.bins}{if x and y are a mix of continuous and discrete variables a discretization is computed
in the continuous one. This parameter indicates the number of intervals to use in the discretization. The default value is 3.}
}
\value{
A matrix containing the correlation between attribute pairs.
}
\description{
This function returns the correlation matrix between the \code{\linkS4class{Attribute}} pairs of the \code{\linkS4class{DataSet}}
}
\examples{
 ds<-dataset(list(attribute(c(2,3,6,4)),attribute(c(1.2,2.3,4.5,7.8))),"Data")
corr<-correlation(ds)
correlation(ds,"EF",4)
correlation(ds,num.bins=4)
correlation(ds,discretizationType="EF")

}
