% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{class}
\name{DataSet-class}
\alias{DataSet-class}
\title{An S4 class to represent a \code{\linkS4class{Dataset}}}
\description{
An S4 class to represent a \code{\linkS4class{Dataset}}
}
\section{Slots}{

\describe{
\item{\code{size}}{Numeric value indicating the size of the \code{\linkS4class{Dataset}}}

\item{\code{data}}{List containing \code{\linkS4class{Attribute}} objects}

\item{\code{name}}{Character with the name of the \code{\linkS4class{Dataset}}}
}}

