% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{discretize,DataSet-method}
\alias{discretize,DataSet-method}
\title{Function to discretize a given \code{\linkS4class{DataSet}}}
\usage{
\S4method{discretize}{DataSet}(x, num.bins, type, columns)
}
\arguments{
\item{x}{\code{\linkS4class{DataSet}} class vector}

\item{num.bins}{Numeric value indicating the number of intervals. Default: half the length of the data}

\item{type}{a Character indicating the type of discretization: Default "EW"(Equal Width) or "EF"(Equal Frequency)}

\item{columns}{Numeric vector indicating the columns in which the discretization must be applied. By default the discretization of every column will be computed}
}
\value{
A vector containing the entropy of each column data if the data is discrete. NA otherwise
}
\description{
This function computes the dicretization of a given \code{\linkS4class{DataSet}}
}
\examples{
ds<-dataset(list(attribute(c(2,3,6,4)),attribute(c(1.2,2.3,4.5,7.8))),"Data")
dsDisc<-discretize(ds,5,"EW")
dsDisc<-discretize(ds,5,"EW",columns=c(1))
dsDisc<-discretize(ds,type="EF")
}
