% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{discretize}
\alias{discretize}
\title{Function to apply discretization to an Attribute}
\usage{
discretize(x, num.bins, type, columns)
}
\arguments{
\item{x}{an attribute composed by real numbers}

\item{num.bins}{Numeric value indicating the number of intervals. Default: half the length of the data}

\item{type}{a Character indicating the type of discretization: Default "EW"(Equal Width) or "EF"(Equal Frequency)}
}
\value{
An Attribute with the factor containing the equal width discretization
}
\description{
This function applies equal width discretization to an Attribute
}
