% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discretizeEF}
\alias{discretizeEF}
\title{Function to apply equal frequency discretization to a vector}
\usage{
discretizeEF(x, num.bins)
}
\arguments{
\item{x}{a vector composed by real numbers}

\item{num.bins}{number of intervals}
}
\value{
A factor with the equal frequency discretization
}
\description{
This function applies equal frequency discretization to a vector
}
