% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discretizeEW}
\alias{discretizeEW}
\title{Function to apply equal width discretization to a vector}
\usage{
discretizeEW(x, num.bins)
}
\arguments{
\item{x}{a vector composed by real numbers}

\item{num.bins}{number of intervals}
}
\value{
A factor with the equal width discretization
}
\description{
This function applies equal width discretization to a vector
}
