% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discretizePoints}
\alias{discretizePoints}
\title{Function to discretize a vector given a list of cutpoints}
\usage{
discretizePoints(x, cut.points)
}
\arguments{
\item{x}{a vector composed by real numbers}

\item{cut.points}{a list with the points at which the vector has to be cut}
}
\value{
A factor with the discretization
}
\description{
This function dicretizes a vector given the list of points to cut
}
