% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{entropy,DataSet-method}
\alias{entropy,DataSet-method}
\title{Function to compute the entropy of a given \code{\linkS4class{DataSet}}}
\usage{
\S4method{entropy}{DataSet}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{DataSet}} class vector}
}
\value{
A vector containing the entropy of each column data if the data is discrete. NA otherwise
}
\description{
This function computes the entropy of a given \code{\linkS4class{DataSet}}
}
\examples{
attr<-attribute(c(0,1,2,0,1,2))
attr<-discretize(attr,3,"EF")
ds<-dataset(c(attr,attribute(c(1.2,2.3,4.5,7.8))),"Data")
entropy(ds)
entropy(attr)
}
