% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{entropy}
\alias{entropy}
\title{Function to compute the entropy of a given Attribute}
\usage{
entropy(x)
}
\arguments{
\item{x}{a vector composed by discrete variables}
}
\value{
A real number
}
\description{
This function computes the entropy of a given vector
}
\examples{
attr<-attribute(c(0,1,2,0,1,2))
attr<-discretize(attr,3,"EF")
entropy(attr)

}
