% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{filter}
\alias{filter}
\title{Function to filter the \code{\linkS4class{DataSet}} attributes}
\usage{
filter(x, threshold, FUN, inverse, columns)
}
\arguments{
\item{x}{\code{\linkS4class{DataSet}} class vector}

\item{threshold}{numeric value that indicates the limit from which to remove the attribute}

\item{FUN}{function by which filter the data, it has to return a value per attribute. FUN=correlation by default}

\item{inverse}{If TRUE the attribute has to be below the threshold to remove it.By default FALSE}

\item{columns}{columns to apply the function to}
}
\value{
A \code{\linkS4class{DataSet}} without the filtered attributes
}
\description{
This function returns the filtered \code{\linkS4class{DataSet}} without the unnecessary attributes
}
\examples{
ds<-dataset(list(attribute(c(2,3,6,4)),attribute(c(1.2,2.3,4.5,7.8))),"Data")
filter(ds,0.4)
}
