% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{getName}
\alias{getName}
\title{Function to get the name of the Attribute}
\usage{
getName(x)
}
\arguments{
\item{x}{an Attribute class vector}
}
\value{
A character containing the name of the attribute.
}
\description{
This function returns the names of the columns of a given DataSet
}
\examples{
attr<- attribute(c(2.3,5.8,4.3,3.2),name="Mark")
getName(attr)

}
