% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{getVector}
\alias{getVector}
\title{Function to get the vector of the Attribute}
\usage{
getVector(attr)
}
\arguments{
\item{attr}{an Attribute class vector}
}
\value{
A vector containing the information of the attribute.
}
\description{
This function returns the vector of a given Attribute
}
\examples{
attr<- attribute(c(2.3,5.8,4.3,3.2),name="Mark")
getVector(attr)

}
