% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadDataSet}
\alias{loadDataSet}
\title{Function to read a CSV file and save it into a \code{\linkS4class{DataSet}}}
\usage{
loadDataSet(path, header, sep)
}
\arguments{
\item{path}{a character that contains a path to the CSV file}

\item{header}{logical indicating if the first row of data corresponds to the names of the \code{\linkS4class{Attribute}}. Default value is FALSE.}

\item{sep}{the character separator of the data. Default value is ",".}
}
\value{
A DataSet containing the data of the CSV file.
}
\description{
This function loads a CSV file into a \code{\linkS4class{DataSet}} object
}
