% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{normalize,DataSet-method}
\alias{normalize,DataSet-method}
\title{Function to normalize a \code{\linkS4class{DataSet}}}
\usage{
\S4method{normalize}{DataSet}(x, columns)
}
\arguments{
\item{x}{a \code{\linkS4class{DataSet}} class vector}

\item{columns}{a vector indicating the columns to which the normalization has to be applied. By default it will be applied to every columns}
}
\value{
A normalized \code{\linkS4class{DataSet}}
}
\description{
This function normalizes a given \code{\linkS4class{DataSet}}
}
\examples{
data<-dataset(c(attribute(c(2,3,6,4)),attribute(c(2,7,8,9))))
normalizedData<- normalize(data)
normalize(data,c(1))
}
