% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{normalize}
\alias{normalize}
\title{Function to normalize a given Attribute}
\usage{
normalize(x, columns)
}
\arguments{
\item{x}{an Attribute class vector}
}
\value{
A normalized attribute
}
\description{
This function normalizes a given vector
}
\examples{
attr<- attribute(c(2.3,5.8,4.3,3.2))
normalize(attr)

}
