% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{rocAuc}
\alias{rocAuc}
\title{Function to compute the AUC-ROC of the DataSet}
\usage{
rocAuc(dat, vIndex, classIndex)
}
\arguments{
\item{dat}{DataSet class vector}

\item{vIndex}{index of the continuous variable to compute the AUC}

\item{classIndex}{index of the class}
}
\value{
The value of the AUC-ROC
}
\description{
This function computes the Area Under the Curve ROC of the DataSet
}
\examples{
attr1<-attribute(c(1,0,1,0),name="Class")
attr2<-attribute(c(3.1,7.8,5.6,4.3),name="Class")
ds<-dataset(list(attr1))
ds<-addAttribute(ds,attr2)
rocAuc(ds,1,2)
}
