% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{rocPlot}
\alias{rocPlot}
\title{Function to visualize the ROC curve of a given DataSet}
\usage{
rocPlot(dat, vIndex, classIndex)
}
\arguments{
\item{dat}{DataSet class vector}

\item{vIndex}{index of the continuous variable to compute the AUC}

\item{classIndex}{index of the class}
}
\value{
A plot with the curve
}
\description{
This function visualizes the Curve ROC of the DataSet
}
\examples{
attr1<-attribute(c(1,0,1,0),name="Class")
attr2<-attribute(c(3.1,7.8,5.6,4.3))
attr3<-attribute(c(3,7,5,7))
ds<-dataset(list(attr1))
ds<-addAttribute(ds,attr2)
ds<-addAttribute(ds,attr3)
rocPlot(ds,2,1)

}
