% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\docType{methods}
\name{show,Attribute-method}
\alias{show,Attribute-method}
\title{Function to show the Attribute}
\usage{
\S4method{show}{Attribute}(object)
}
\arguments{
\item{object}{Attribute class object}
}
\value{
A character object with the information of the Attribute
}
\description{
This function shows the Attribute
}
\examples{
attr<- attribute(c(2.3,5.8,4.3,3.2),name="Mark")
attr
}
