% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{show,DataSet-method}
\alias{show,DataSet-method}
\title{Function to print the \code{\linkS4class{DataSet}}}
\usage{
\S4method{show}{DataSet}(object)
}
\arguments{
\item{object}{\code{\linkS4class{DataSet}} class vector}
}
\value{
Shows the information of the \code{\linkS4class{DataSet}}.
}
\description{
This function prints the \code{\linkS4class{DataSet}}
}
\examples{
ds<-dataset(list(attribute(c(2,3,6,4)),attribute(c(1.2,2.3,4.5,7.8))),"Data")
ds

}
