% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{standardize}
\alias{standardize}
\title{Function to standardize a given Attribute}
\usage{
standardize(x, columns)
}
\arguments{
\item{x}{an Attribute class vector}

\item{columns}{columns that need to be standardized}
}
\value{
A standardized attribute
}
\description{
This function standardizes a given vector
}
\examples{
attr<- attribute(c(2.3,5.8,4.3,3.2))
standardize(attr)

}
