% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\docType{methods}
\name{variance,DataSet-method}
\alias{variance,DataSet-method}
\title{Function to compute the variance of a given \code{\linkS4class{DataSet}}}
\usage{
\S4method{variance}{DataSet}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{DataSet}} Object}
}
\value{
A vector containing the variance of each column data. If an \code{\linkS4class{Attribute}} is categorical it will return NA
}
\description{
This function computes the variance of a given \code{\linkS4class{DataSet}}
}
\examples{
ds<-dataset(list(attribute(c(2,3,6,4)),attribute(c(1.2,2.3,4.5,7.8))),"Data")
variance(ds)

}
