% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{variance}
\alias{variance}
\title{Function to compute the variance of a given Attribute}
\usage{
variance(x)
}
\arguments{
\item{x}{an Attribute class vector}
}
\value{
A real number containing the variance of the attribute
}
\description{
This function computes the variance of a given Attribute
}
\examples{
attr<- attribute(c(2.3,5.8,4.3,3.2))
variance(attr)

}
