% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeDataSet}
\alias{writeDataSet}
\title{Function to write a CSV file with a \code{\linkS4class{DataSet}} information}
\usage{
writeDataSet(data, file.path)
}
\arguments{
\item{data}{the \code{\linkS4class{DataSet}} object to save}

\item{file.path}{a character with the path where the \code{\linkS4class{DataSet}} has to be saved}
}
\value{
It creates a file in the given file.path containing the data of the \code{\linkS4class{DataSet}} object.
}
\description{
This function writes a CSV file with a \code{\linkS4class{DataSet}} information
}
