% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeFile}
\alias{writeFile}
\title{Function to save a log text into a file}
\usage{
writeFile(text, file.path)
}
\arguments{
\item{text}{a character that contains the text to write}

\item{file.path}{a character with the path where the text has to be saved}
}
\description{
This function saves the given text with the actual time and date in a file in the given path
}
